+++ src/components/DropTarget/index.js
@@ -1,9 +1,7 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 
 import './style.css';
 
 export default () => (
-	<Fragment>
-		<div className="dropzone-placeholder" />
-	</Fragment>
+	<div className="dropzone-placeholder" />
 );
