+++ src/containers/App/index.js
@@ -1,12 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import DropToUpload from 'react-drop-to-upload';
 import { connect } from 'react-redux';
 
 import SkewerContainer from '../Skewer';
-import DropTarget from '../../components/DropTarget';
-import { setImage } from '../../store/actions';
-import { readAsImage } from '../../helpers';
+import DropTargetContainer from '../DropTarget';
 import DropZone from '../DropZone';
 
 import './style.css';
@@ -15,19 +12,7 @@ const mapStateToProps = ( state ) => ( {
 	imageLoaded: ! ! state.image.src,
 } );
 
-const mapDispatchToProps = ( dispatch ) => ( {
-	onDrop( acceptedFiles ) {
-		readAsImage( acceptedFiles[ 0 ] ).then( image => dispatch( setImage( image ) ) );
-	},
-	onFileChange( evt ) {
-		readAsImage( evt.target.files[0] ).then( image => dispatch( setImage( image ) ) );
-	},
-} );
-
-const App = ( {
-	imageLoaded,
-	onFileChange,
-} ) => (
+export const App = ( { imageLoaded } ) => (
 	<DropZone>
 		<div className="App">
 			{ imageLoaded ? (
@@ -36,12 +21,7 @@ const App = ( {
 				<header className="App-header">
 					<h1>Image Kebab<em>!</em></h1>
 
-						<DropTarget />
-						<input
-							className="image-upload-button"
-							type="file"
-							onChange={ onFileChange }
-						/>
+					<DropTargetContainer />
 
 					<p>Drag-and-drop the image to skewer into this browser window.</p>
 
@@ -58,4 +38,4 @@ App.propTypes = {
 	imageLoaded: PropTypes.bool.isRequired,
 };
 
-export default connect( mapStateToProps, mapDispatchToProps )( App );
+export default connect( mapStateToProps )( App );
